+++ src/Core/UI/Overlay.cpp
@@ -71,7 +71,7 @@ namespace GuGu {
 			AlignmentArrangeResult xResult = AlignChild<Orientation::Horizontal>(curChild, allocatedGeometry.getLocalSize().x);
 			AlignmentArrangeResult yResult = AlignChild<Orientation::Vertical>(curChild, allocatedGeometry.getLocalSize().y);
 
-			arrangedWidgetArray.pushWidget(allocatedGeometry.getChildGeometry(math::double2(xResult.m_size, yResult.m_size), math::double2(xResult.m_offset, yResult.m_offset), allocatedGeometry.getAccumulateTransform()), curChild.getChildWidget());
+			arrangedWidgetArray.pushWidget(allocatedGeometry.getChildGeometry(math::float2(xResult.m_size, yResult.m_size), math::float2(xResult.m_offset, yResult.m_offset), allocatedGeometry.getAccumulateTransform()), curChild.getChildWidget());
 		}
 	}
 	SlotBase* Overlay::getSlot(uint32_t index)
